basic-template/lambdas/zendesk/requirements.txt
phidata
openai
boto3
requests